Removed options -FSx, -RSx. Per-file K=V options.
Now you can set the options NF, FS and RS for each file using the
syntax KEY=VALUE on the command line before the filename. This
should be more convenient for the user than the previous
approach of setting the options in bulk. It should make it easier
to add new options to the program.

Moved most of the loop in ::sqawk::script::main to ::sqawk::sqawk
method read-file.