tests.tcl: Split large tests into multiple smaller
Tests that accumulate results of multiple commands in a result
list have proved to be an antipattern. Splitting them into sets
of smaller tests makes the purpose of each individual test easier
to understand, lets you compare the actual and the desired output
more quickly and reduces the need for cleanup scripts.