lib/classes/table.tcl: Fix stale insert stmt use
In the method insert-rows the value of curNF is not updated when
reusing a cached insert statement. This leads to fields not
being inserted into the table if you have one or more records
of N fields (a new statement for N gets generated at this point),
then M > N fields (a new statement for M), then N again (the
cached statement for N is loaded), then M fields again (the cached
statement for M isn't loaded).